+++ modules/caption/gemini.py
@@ -11,7 +11,7 @@ class GoogleGeminiPipeline():
     def __init__(self, model_name: str):
         self.model = model_name.split(' (')[0]
         from installer import install
-        install('google-genai==1.52.0')
+        install('google-genai==2.22.0')
         from google import genai # pylint: disable=no-name-in-module
         args = self.get_args()
         self.client = genai.Client(**args)